added basic functionality in qml
